Fixed the bug in mark calculation.
We have to add online help, fix the UI and we're done!

+++ PythonGuru/app/src/main/java/gr/example/zografos/vasileios/pythonguru/ProgressActivity.kt
@@ -45,7 +45,7 @@ class ProgressActivity : AppCompatActivity() {
 
         if (!marksStr.isEmpty()) {
             val marksStrs = marksStr.split(",")
-            val realMarks = marksStrs.map { markStr -> markStr.toFloat() }
+            val realMarks = marksStrs.map { markStr -> markStr.toDouble() }
 
             // display marks and comments
             marksViews.forEachIndexed { index, markView ->
